[core] Consider tile overscaledZ when querying rendered features
Previously, the ScreenBox represented abstract points in the
coordinate space. However, the query rendered features method
works at the level of tiles and therefore at an actual an implied
zoom level. The ScreenBox passed in from the SDK should represent
the query rect in points and that rect should be adjusted to cover
the per zoom level scale.

This commit passes the ScreenBox down to the source's
query `queryRenderedFeatures` method so that it can be adjusted
for each evaluated tile. With this change, the expect number of
point annotations are returned for any arbitrary ScreenBox rect (in
points) at any zoom level.