[ios, macos] Refactor polygon feature to return mbgl feature
MGLPolygon now returns its geometry wrapped in an mbgl feature object
that is then given an mbgl propery_map. The property map translation
is just a stub for now and uses a copy of the logic in NSExpression+
MGLAdditions. It does n ot handle recursiveness in the attributes
structure. This change does allow us to handle several types of
GeoJSON options / attributes and load an mbgl geojson source with
an mbgl feature collection without doing a roundtrip through
NSJSONSerialization.